Dispatch reporters via a name→fn map; fold row mapping into json reporter
- cli/reporters/mod.ts exports a `reporters` map (pretty/json/ndjson → fn) and
  derives `ReporterName` from its keys, so `check` dispatches with a lookup
  instead of an if/else chain. json/ndjson share one reporter, differing only by
  the `format` passed.
- The row mapping (toReportRow + ReportRow) was only ever used by the json
  reporter, so it moves into json.ts (with its inline tests) rather than living
  as a standalone reporters/reportRow.ts module.
